Removed duplicated file
